+++ Calculations/Inputs.cs
@@ -19,19 +19,19 @@ public static partial class Calculations
     /// <returns></returns>
     public static StockData CalculateTypicalPrice(this StockData stockData)
     {
-        List<decimal> tpList = new();
+        List<double> tpList = new();
         List<Signal> signalsList = new();
         var (inputList, highList, lowList, _, _) = GetInputValuesList(stockData);
 
         for (int i = 0; i < stockData.Count; i++)
         {
-            decimal currentHigh = highList[i];
-            decimal currentLow = lowList[i];
-            decimal currentClose = inputList[i];
-            decimal prevTypicalPrice1 = i >= 1 ? tpList[i - 1] : 0;
-            decimal prevTypicalPrice2 = i >= 2 ? tpList[i - 2] : 0;
+            double currentHigh = highList[i];
+            double currentLow = lowList[i];
+            double currentClose = inputList[i];
+            double prevTypicalPrice1 = i >= 1 ? tpList[i - 1] : 0;
+            double prevTypicalPrice2 = i >= 2 ? tpList[i - 2] : 0;
 
-            decimal typicalPrice = (currentHigh + currentLow + currentClose) / 3;
+            double typicalPrice = (currentHigh + currentLow + currentClose) / 3;
             tpList.AddRounded(typicalPrice);
 
             Signal signal = GetCompareSignal(typicalPrice - prevTypicalPrice1, prevTypicalPrice1 - prevTypicalPrice2);
@@ -56,18 +56,18 @@ public static partial class Calculations
     /// <returns></returns>
     public static StockData CalculateMedianPrice(this StockData stockData)
     {
-        List<decimal> medianPriceList = new();
+        List<double> medianPriceList = new();
         List<Signal> signalsList = new();
         var (_, highList, lowList, _, _) = GetInputValuesList(stockData);
 
         for (int i = 0; i < stockData.Count; i++)
         {
-            decimal currentLow = lowList[i];
-            decimal currentHigh = highList[i];
-            decimal prevMedianPrice1 = i >= 1 ? medianPriceList[i - 1] : 0;
-            decimal prevMedianPrice2 = i >= 2 ? medianPriceList[i - 2] : 0;
+            double currentLow = lowList[i];
+            double currentHigh = highList[i];
+            double prevMedianPrice1 = i >= 1 ? medianPriceList[i - 1] : 0;
+            double prevMedianPrice2 = i >= 2 ? medianPriceList[i - 2] : 0;
 
-            decimal medianPrice = (currentHigh + currentLow) / 2;
+            double medianPrice = (currentHigh + currentLow) / 2;
             medianPriceList.AddRounded(medianPrice);
 
             Signal signal = GetCompareSignal(medianPrice - prevMedianPrice1, prevMedianPrice1 - prevMedianPrice2);
@@ -92,20 +92,20 @@ public static partial class Calculations
     /// <returns></returns>
     public static StockData CalculateFullTypicalPrice(this StockData stockData)
     {
-        List<decimal> fullTpList = new();
+        List<double> fullTpList = new();
         List<Signal> signalsList = new();
         var (inputList, highList, lowList, openList, _) = GetInputValuesList(stockData);
 
         for (int i = 0; i < stockData.Count; i++)
         {
-            decimal currentHigh = highList[i];
-            decimal currentLow = lowList[i];
-            decimal currentClose = inputList[i];
-            decimal currentOpen = openList[i];
-            decimal prevTypicalPrice1 = i >= 1 ? fullTpList[i - 1] : 0;
-            decimal prevTypicalPrice2 = i >= 2 ? fullTpList[i - 2] : 0;
+            double currentHigh = highList[i];
+            double currentLow = lowList[i];
+            double currentClose = inputList[i];
+            double currentOpen = openList[i];
+            double prevTypicalPrice1 = i >= 1 ? fullTpList[i - 1] : 0;
+            double prevTypicalPrice2 = i >= 2 ? fullTpList[i - 2] : 0;
 
-            decimal typicalPrice = (currentHigh + currentLow + currentClose + currentOpen) / 4;
+            double typicalPrice = (currentHigh + currentLow + currentClose + currentOpen) / 4;
             fullTpList.AddRounded(typicalPrice);
 
             Signal signal = GetCompareSignal(typicalPrice - prevTypicalPrice1, prevTypicalPrice1 - prevTypicalPrice2);
@@ -130,19 +130,19 @@ public static partial class Calculations
     /// <returns></returns>
     public static StockData CalculateWeightedClose(this StockData stockData)
     {
-        List<decimal> weightedCloseList = new();
+        List<double> weightedCloseList = new();
         List<Signal> signalsList = new();
         var (inputList, highList, lowList, _, _) = GetInputValuesList(stockData);
 
         for (int i = 0; i < stockData.Count; i++)
         {
-            decimal currentHigh = highList[i];
-            decimal currentLow = lowList[i];
-            decimal currentClose = inputList[i];
-            decimal prevClose = i >= 1 ? inputList[i - 1] : 0;
+            double currentHigh = highList[i];
+            double currentLow = lowList[i];
+            double currentClose = inputList[i];
+            double prevClose = i >= 1 ? inputList[i - 1] : 0;
 
-            decimal prevWeightedClose = weightedCloseList.LastOrDefault();
-            decimal weightedClose = (currentHigh + currentLow + (currentClose * 2)) / 4;
+            double prevWeightedClose = weightedCloseList.LastOrDefault();
+            double weightedClose = (currentHigh + currentLow + (currentClose * 2)) / 4;
             weightedCloseList.AddRounded(weightedClose);
 
             Signal signal = GetCompareSignal(currentClose - weightedClose, prevClose - prevWeightedClose);
@@ -168,20 +168,20 @@ public static partial class Calculations
     /// <returns></returns>
     public static StockData CalculateMidpoint(this StockData stockData, int length = 14)
     {
-        List<decimal> midpointList = new();
+        List<double> midpointList = new();
         List<Signal> signalsList = new();
         var (inputList, _, _, _, _) = GetInputValuesList(stockData);
         var (highestList, lowestList) = GetMaxAndMinValuesList(inputList, length);
 
         for (int i = 0; i < stockData.Count; i++)
         {
-            decimal currentValue = inputList[i];
-            decimal prevValue = i >= 1 ? inputList[i - 1] : 0;
-            decimal highest = highestList[i];
-            decimal lowest = lowestList[i];
+            double currentValue = inputList[i];
+            double prevValue = i >= 1 ? inputList[i - 1] : 0;
+            double highest = highestList[i];
+            double lowest = lowestList[i];
 
-            decimal prevMidPoint = midpointList.LastOrDefault();
-            decimal midpoint = (highest + lowest) / 2;
+            double prevMidPoint = midpointList.LastOrDefault();
+            double midpoint = (highest + lowest) / 2;
             midpointList.AddRounded(midpoint);
 
             var signal = GetCompareSignal(currentValue - midpoint, prevValue - prevMidPoint);
@@ -207,20 +207,20 @@ public static partial class Calculations
     /// <returns></returns>
     public static StockData CalculateMidprice(this StockData stockData, int length = 14)
     {
-        List<decimal> midpriceList = new();
+        List<double> midpriceList = new();
         List<Signal> signalsList = new();
         var (inputList, highList, lowList, _, _) = GetInputValuesList(stockData);
         var (highestList, lowestList) = GetMaxAndMinValuesList(highList, lowList, length);
 
         for (int i = 0; i < stockData.Count; i++)
         {
-            decimal currentValue = inputList[i];
-            decimal prevValue = i >= 1 ? inputList[i - 1] : 0;
-            decimal highest = highestList[i];
-            decimal lowest = lowestList[i];
+            double currentValue = inputList[i];
+            double prevValue = i >= 1 ? inputList[i - 1] : 0;
+            double highest = highestList[i];
+            double lowest = lowestList[i];
 
-            decimal prevMidPrice = midpriceList.LastOrDefault();
-            decimal midPrice = (highest + lowest) / 2;
+            double prevMidPrice = midpriceList.LastOrDefault();
+            double midPrice = (highest + lowest) / 2;
             midpriceList.AddRounded(midPrice);
 
             var signal = GetCompareSignal(currentValue - midPrice, prevValue - prevMidPrice);
